requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
pytest
pytest-asyncio
httpx